feat(#369): SD3-Medium full txt2img MLX pipeline + image_gen wiring
Native from-scratch MLX port of Stable Diffusion 3 Medium txt2img under
fusion_mlx/image/sd3/. Completes the pipeline started in prior commits
(MMDiT transformer, scheduler, VAE, CLIP text model).

- generate.py: SD3Pipeline loads transformer+VAE from argmaxinc checkpoint
  and CLIP-L/CLIP-G/T5-XXL from frankjoshua diffusers repo. Reuses mflux
  CLIPEncoder (CLIP-L) and T5Encoder (T5-XXL); CLIP-G via the parametrized
  CLIPTextModel. Encodes prompt (pooled = CLIP-L ⊕ CLIP-G EOS, context =
  T5), runs FlowMatchEuler rectified-flow denoise loop with CFG, decodes
  with VAE -> PIL. Tokenizers loaded from separate tokenizer/ tokenizer_2/
  tokenizer_3/ folders.
- config.py: SD3ModelPaths gains tokenizer subfolders; ClipGConfig uses
  20 heads (head_dim=64) per StabilityAI diffusers config.
- text_encoder.py: cast causal mask to hidden.dtype for fp16 sdpa.
- engines/image_gen.py: VARIANT_MAP["sd3"] -> SD3Pipeline; _infer_variant
  recognizes sd3/stable-diffusion-3; native variants skip mflux
  ModelConfig factory; negative_prompt + shift forwarded for SD3.
- tests/unit/test_image_gen_sd3.py: config, variant inference, T5 weight
  remap, CLIPTextModel param names/pooled shape, pipeline construction.
- docs/sd3-image.md + README mention.

Component-verified against real argmaxinc/mlx-stable-diffusion-3-medium
checkpoint: transformer 491/491 params, VAE 244/244 params, CLIP-L real
weights forward finite, MMDiT + VAE synthetic forward finite, tokenizers
load. Full real-model E2E pending text-encoder download.

+++ fusion_mlx/image/sd3/text_encoder.py
@@ -123,7 +123,7 @@ class CLIPTextModel(nn.Module):
 
     def __call__(self, tokens: mx.array) -> mx.array:
         hidden = self.text_model.embeddings(tokens)
-        mask = _causal_mask(hidden.shape)
+        mask = _causal_mask(hidden.shape).astype(hidden.dtype)
         hidden = self.text_model.encoder(hidden, mask)
         hidden = self.text_model.final_layer_norm(hidden)
         eos = mx.argmax(tokens, axis=-1)
